Algumas correções

+++ Exercícios/ex029.py
@@ -1,4 +1,4 @@
-vel = int(input('Tell me the speed (Km/h): '))
+vel = float(input('Tell me the speed (Km/h): '))
 if vel > 80:
     price = (vel - 80) * 7
     print('FINED, YOU OWN ME ${}'.format(price))
